resort software code database [WIP] - split av_controller.c into several sub-files - WorkInProgress = still needs refactoring
