+++ software/sys_controller/tvp7002/tvp7002.c
@@ -38,6 +38,11 @@ const ypbpr_to_rgb_csc_t csc_coeffs[] = {
     { "Rec. 709", 0x2000, 0x0000, 0x323E, 0x2000, 0xFA04, 0xF113, 0x2000, 0x3B61, 0x0000 },    // eq. 105
 };
 
+const alt_u8 Kvco[] = {75, 85, 150, 200};
+const char const *Kvco_str[] = { "Ultra low", "Low", "Medium", "High" };
+
+const alt_u32 clkrate[] = {27000000, 6500000}; //in MHz
+
 extern mode_data_t video_modes[];
 
 static inline void tvp_set_ssthold(alt_u8 vsdetect_thold)
@@ -111,7 +116,7 @@ void tvp_set_alc(alt_u8 disable_alc, video_type type)
 }
 
 
-inline alt_u32 tvp_readreg(alt_u32 regaddr)
+alt_u32 tvp_readreg(alt_u32 regaddr)
 {
     I2C_start(I2CA_BASE, TVP_BASE, 0);
     I2C_write(I2CA_BASE, regaddr, 1);   //don't use repeated start as it seems unreliable at 400kHz
@@ -119,14 +124,14 @@ inline alt_u32 tvp_readreg(alt_u32 regaddr)
     return I2C_read(I2CA_BASE,1);
 }
 
-inline void tvp_writereg(alt_u32 regaddr, alt_u8 data)
+void tvp_writereg(alt_u32 regaddr, alt_u8 data)
 {
     I2C_start(I2CA_BASE, TVP_BASE, 0);
     I2C_write(I2CA_BASE, regaddr, 0);
     I2C_write(I2CA_BASE, data, 1);
 }
 
-inline void tvp_reset()
+void tvp_reset()
 {
     usleep(10000);
     IOWR_ALTERA_AVALON_PIO_DATA(PIO_0_BASE, 0x00);
@@ -135,7 +140,7 @@ inline void tvp_reset()
     usleep(10000);
 }
 
-inline void tvp_disable_output()
+void tvp_disable_output()
 {
     usleep(10000);
     tvp_writereg(TVP_MISCCTRL1, 0x13);
@@ -144,7 +149,7 @@ inline void tvp_disable_output()
     usleep(10000);
 }
 
-inline void tvp_enable_output()
+void tvp_enable_output()
 {
     usleep(10000);
     tvp_writereg(TVP_MISCCTRL1, 0x11);
@@ -153,7 +158,7 @@ inline void tvp_enable_output()
     usleep(10000);
 }
 
-inline void tvp_set_hpllcoast(alt_u8 pre, alt_u8 post)
+void tvp_set_hpllcoast(alt_u8 pre, alt_u8 post)
 {
     tvp_writereg(TVP_HPLLPRECOAST, pre);
     tvp_writereg(TVP_HPLLPOSTCOAST, post);
